Reorganize main services
