+++ kraftwerk-api/pom.xml
@@ -1,39 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
- 
-  <parent>
-    <groupId>fr.insee.kraftwerk</groupId>
-    <artifactId>kraftwerk</artifactId>
-    <version>1.3.1</version>
-  </parent>
-  <artifactId>kraftwerk-api</artifactId>
- 
-  <packaging>jar</packaging> 
-  
-	<dependencies>
+	<modelVersion>4.0.0</modelVersion>
 	
+	<parent>
+		<groupId>fr.insee.kraftwerk</groupId>
+		<artifactId>kraftwerk</artifactId>
+		<version>1.3.1</version>
+	</parent>
+	
+	<artifactId>kraftwerk-api</artifactId>
+	<packaging>jar</packaging>
+	<name>kraftwerk-api</name>
+	
+	<properties>
+		<kraftwerk.version>1.3.1</kraftwerk.version>
+	</properties>
+	
+	<dependencies>
 		<!-- Kraftwerk Core -->
 		<dependency>
-		    <groupId>fr.insee.kraftwerk</groupId>
-		    <artifactId>kraftwerk-core</artifactId>
-		    <version>1.3.1</version>
-	    </dependency>
-
+			<groupId>fr.insee.kraftwerk</groupId>
+			<artifactId>kraftwerk-core</artifactId>
+			<version>${kraftwerk.version}</version>
+		</dependency>
 		<!-- Spring -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-		
 		<!-- SWAGGER -->
-		 <dependency>
-		      <groupId>org.springdoc</groupId>
-		      <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
-		      <version>2.0.2</version>
-		 </dependency>
-
+		<dependency>
+			<groupId>org.springdoc</groupId>
+			<artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+			<version>2.1.0</version>
+		</dependency>
 	</dependencies>
-
 	<build>
 		<plugins>
 			<plugin>
