+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/configuration/OpenApiConfiguration.java
@@ -1,4 +1,4 @@
-package fr.insee.kraftwerk.api;
+package fr.insee.kraftwerk.api.configuration;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
