+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/process/MainProcessing.java
@@ -1,0 +1,173 @@
+package fr.insee.kraftwerk.api.process;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
+
+import org.springframework.beans.factory.annotation.Value;
+
+import fr.insee.kraftwerk.core.KraftwerkError;
+import fr.insee.kraftwerk.core.exceptions.KraftwerkException;
+import fr.insee.kraftwerk.core.exceptions.NullException;
+import fr.insee.kraftwerk.core.inputs.ModeInputs;
+import fr.insee.kraftwerk.core.inputs.UserInputs;
+import fr.insee.kraftwerk.core.metadata.MetadataUtils;
+import fr.insee.kraftwerk.core.metadata.VariablesMap;
+import fr.insee.kraftwerk.core.sequence.BuildBindingsSequence;
+import fr.insee.kraftwerk.core.sequence.ControlInputSequence;
+import fr.insee.kraftwerk.core.sequence.MultimodalSequence;
+import fr.insee.kraftwerk.core.sequence.UnimodalSequence;
+import fr.insee.kraftwerk.core.sequence.WriterSequence;
+import fr.insee.kraftwerk.core.utils.TextFileWriter;
+import fr.insee.kraftwerk.core.vtl.VtlBindings;
+import jakarta.annotation.PostConstruct;
+import lombok.extern.log4j.Log4j2;
+
+@Log4j2
+public class MainProcessing {
+
+	@Value("${fr.insee.postcollecte.files}")
+	private String defaultDirectory;
+
+	private ControlInputSequence controlInputSequence;
+	private boolean fileByFile;
+
+	/* SPECIFIC VARIABLES */
+	private String inDirectoryParam;
+	private Path inDirectory;
+	private UserInputs userInputs; // if main is called on all files
+	List<UserInputs> userInputsList; // for file by file process
+	private VtlBindings vtlBindings = new VtlBindings();
+	private List<KraftwerkError> errors = new ArrayList<>();
+	private Map<String, VariablesMap> metadataVariables;
+
+	public MainProcessing(String inDirectoryParam, boolean fileByFile) {
+		super();
+		this.inDirectoryParam = inDirectoryParam;
+		this.fileByFile = fileByFile;
+		controlInputSequence = new ControlInputSequence(defaultDirectory);
+	}
+
+	@PostConstruct
+	public void initializeWithProperties() {
+		controlInputSequence = new ControlInputSequence(defaultDirectory);
+	}
+
+	public void runMain() throws KraftwerkException {
+		init();
+		if (Boolean.TRUE.equals(fileByFile)) { //iterate on files
+			for (UserInputs userInputsFile : userInputsList) {
+				userInputs = userInputsFile;
+				vtlBindings = new VtlBindings();
+				unimodalProcess();
+				multimodalProcess();
+				outputFileWriter();
+			}
+		} else {
+			unimodalProcess();
+			multimodalProcess();
+			outputFileWriter();
+		}
+		writeErrors();
+	}
+
+	/* Step 1 : Init */
+	private void init() throws KraftwerkException {
+		inDirectory = controlInputSequence.getInDirectory(inDirectoryParam);
+
+		String campaignName = inDirectory.getFileName().toString();
+		log.info("Kraftwerk main service started for campaign: " + campaignName);
+
+		userInputs = controlInputSequence.getUserInputs(inDirectory);
+		metadataVariables = MetadataUtils.getMetadata(userInputs.getModeInputsMap());
+
+		if (Boolean.TRUE.equals(fileByFile))
+			userInputsList = getUserInputs(userInputs);
+
+	}
+
+	/* Step 2 : unimodal data */
+	private void unimodalProcess() throws NullException {
+		BuildBindingsSequence buildBindingsSequence = new BuildBindingsSequence();
+		for (String dataMode : userInputs.getModeInputsMap().keySet()) {
+			buildBindingsSequence.buildVtlBindings(userInputs, dataMode, vtlBindings, metadataVariables);
+			UnimodalSequence unimodal = new UnimodalSequence();
+			unimodal.unimodalProcessing(userInputs, dataMode, vtlBindings, errors, metadataVariables);
+		}
+	}
+
+	/* Step 3 : multimodal VTL data processing */
+	private void multimodalProcess() {
+		MultimodalSequence multimodalSequence = new MultimodalSequence();
+		multimodalSequence.multimodalProcessing(userInputs, vtlBindings, errors, metadataVariables);
+	}
+
+	/* Step 4 : Write output files */
+	private void outputFileWriter() {
+		WriterSequence writerSequence = new WriterSequence();
+		writerSequence.writeOutputFiles(inDirectory, vtlBindings, userInputs.getModeInputsMap(),
+				userInputs.getMultimodeDatasetName(), metadataVariables, errors);
+	}
+
+	/* Step 5 : Write errors */
+	private void writeErrors() {
+		TextFileWriter.writeErrorsFile(inDirectory, errors);
+	}
+
+	private static List<UserInputs> getUserInputs(UserInputs source) {
+		List<UserInputs> userInputsList = new ArrayList<>();
+		for (String dataMode : source.getModeInputsMap().keySet()) {
+			List<Path> dataFiles = getFilesToProcess(source, dataMode);
+			for (Path dataFile : dataFiles) {
+				UserInputs currentFileInputs = new UserInputs();
+				currentFileInputs.setUserInputFile(source.getUserInputFile());
+				currentFileInputs.setInputDirectory(source.getInputDirectory());
+				currentFileInputs.setVtlReconciliationFile(source.getVtlReconciliationFile());
+				currentFileInputs.setVtlInformationLevelsFile(source.getVtlInformationLevelsFile());
+				currentFileInputs.setVtlTransformationsFile(source.getVtlTransformationsFile());
+				currentFileInputs.setMultimodeDatasetName(source.getMultimodeDatasetName());
+				ModeInputs sourceModeInputs = source.getModeInputs(dataMode);
+				ModeInputs currentFileModeInputs = new ModeInputs();
+				currentFileModeInputs.setDataFile(dataFile);
+				currentFileModeInputs.setDdiUrl(sourceModeInputs.getDdiUrl());
+				currentFileModeInputs.setLunaticFile(sourceModeInputs.getLunaticFile());
+				currentFileModeInputs.setDataFormat(sourceModeInputs.getDataFormat().toString());
+				currentFileModeInputs.setDataMode(sourceModeInputs.getDataMode());
+				currentFileModeInputs.setModeVtlFile(sourceModeInputs.getModeVtlFile());
+				currentFileModeInputs.setParadataFolder(sourceModeInputs.getParadataFolder());
+				currentFileModeInputs.setReportingDataFile(sourceModeInputs.getReportingDataFile());
+				currentFileInputs.getModeInputsMap().put(dataMode, currentFileModeInputs);
+				userInputsList.add(currentFileInputs);
+			}
+		}
+		return userInputsList;
+	}
+
+	private static List<Path> getFilesToProcess(UserInputs userInputs, String dataMode) {
+		List<Path> files = new ArrayList<>();
+		ModeInputs modeInputs = userInputs.getModeInputs(dataMode);
+		Path dataPath = modeInputs.getDataFile();
+		if (dataPath == null)
+			log.error("Datapath is null");
+		else {
+			if (Files.isRegularFile(dataPath)) {
+				files.add(dataPath);
+			} else if (Files.isDirectory(dataPath)) {
+				try (Stream<Path> stream = Files.list(dataPath)) {
+					stream.forEach(files::add);
+				} catch (IOException e) {
+					log.error(String.format("IOException occurred when trying to list data files of folder: %s",
+							dataPath));
+				}
+			} else {
+				log.warn(String.format("Data path given could not be identified as a file or folder: %s", dataPath));
+			}
+		}
+		return files;
+	}
+
+}
