+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/services/KraftwerkService.java
@@ -1,11 +1,8 @@
-package fr.insee.kraftwerk.api;
+package fr.insee.kraftwerk.api.services;
 
 
 
-import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Path;
-import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
@@ -13,16 +10,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
 import fr.insee.kraftwerk.core.Constants;
-import fr.insee.kraftwerk.core.KraftwerkError;
 import fr.insee.kraftwerk.core.exceptions.KraftwerkException;
 import fr.insee.kraftwerk.core.sequence.ControlInputSequence;
 import fr.insee.kraftwerk.core.utils.FileUtils;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.annotation.PostConstruct;
-import lombok.extern.log4j.Log4j2;
 
-@Log4j2
 @RestController
 @ApiResponses(value = {
 		@ApiResponse(responseCode = "200", description = "Success"),
@@ -50,27 +44,6 @@ public class KraftwerkService {
 		controlInputSequence = new ControlInputSequence(defaultDirectory);
 	}
 	
-
-	protected void writeErrorsFile(Path inDirectory, List<KraftwerkError> errors) {
-		Path tempOutputPath = FileUtils.transformToOut(inDirectory).resolve("errors.txt");
-		FileUtils.createDirectoryIfNotExist(tempOutputPath.getParent());
-
-		//Write errors file
-		if (!errors.isEmpty()) {
-			try (FileWriter myWriter = new FileWriter(tempOutputPath.toFile(),true)){
-				for (KraftwerkError error : errors){
-					myWriter.write(error.toString());
-				}
-				log.info(String.format("Text file: %s successfully written", tempOutputPath));
-			} catch (IOException e) {
-				log.warn(String.format("Error occurred when trying to write text file: %s", tempOutputPath), e);
-			}
-		} else {
-			log.debug("No error found during VTL transformations");
-		}
-	}
-	
-
 	protected ResponseEntity<String> archive(String inDirectoryParam) {
 		Path inDirectory;
 		try {
