+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/services/MainService.java
@@ -1,0 +1,63 @@
+package fr.insee.kraftwerk.api.services;
+
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import fr.insee.kraftwerk.api.process.MainProcessing;
+import fr.insee.kraftwerk.core.exceptions.KraftwerkException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+
+
+
+@RestController
+@Tag(name = "${tag.main}")
+public class MainService extends KraftwerkService {
+
+	@PutMapping(value = "/main")
+	@Operation(operationId = "main", summary = "${summary.main}", description = "${description.main}")
+	public ResponseEntity<String> main(
+			@Parameter(description = "${param.inDirectory}", required = true, example = INDIRECTORY_EXAMPLE) @RequestBody String inDirectoryParam,
+			@Parameter(description = "${param.archiveAtEnd}", required = false) @RequestParam(defaultValue = "false") boolean archiveAtEnd
+			) {
+		MainProcessing mp = new MainProcessing(inDirectoryParam, false);
+		try {
+			mp.runMain();
+		} catch (KraftwerkException e) {
+			return ResponseEntity.status(e.getStatus()).body(e.getMessage());
+		}
+
+		/* Step 4.3- 4.4 : Archive */
+		if (Boolean.TRUE.equals(archiveAtEnd)) archive(inDirectoryParam);
+
+		return ResponseEntity.ok(inDirectoryParam);
+	}
+
+	@PutMapping(value = "/main/file-by-file")
+	@Operation(operationId = "main", summary = "${summary.fileByFile}", description = "${description.fileByFile}")
+	public ResponseEntity<String> mainFileByFile(
+			@Parameter(description = "${param.inDirectory}", required = true, example = INDIRECTORY_EXAMPLE) @RequestBody String inDirectoryParam,
+			@Parameter(description = "${param.archiveAtEnd}", required = false) @RequestParam(defaultValue = "false") boolean archiveAtEnd
+	) {
+		MainProcessing mp = new MainProcessing(inDirectoryParam, true);
+		try {
+			mp.runMain();
+		} catch (KraftwerkException e) {
+			return ResponseEntity.status(e.getStatus()).body(e.getMessage());
+		}
+			/* Step 4.3- 4.4 : Archive */
+		if (Boolean.TRUE.equals(archiveAtEnd)) archive(inDirectoryParam);
+
+		return ResponseEntity.ok(inDirectoryParam);
+	}
+
+
+
+
+}
